Move more NTT-specific network code under network::ntt
Also, the ntt submodule is made private to network. No excuses to
leaking network protocol details.
